requirements.txt
asgiref==3.8.1
attrs==23.2.0
az-iranian-bank-gateways==2.0.5
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-ckeditor==6.7.1
django-js-asset==2.2.0
idna==3.7
isodate==0.6.1
jalalidate==0.1.3
lxml==5.2.2
persiantools==4.2.0
pillow==10.4.0
platformdirs==4.2.2
pycryptodome==3.20.0
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
zeep==4.2.1
